+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 This changelog follows the specifications detailed in: [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
+## Unreleased
+
+### Added
+
+* Added a multinomial version of the random effects alignment function (note this is still not multi-kdma)
+
+### Fixed
+
+* Updated unit tests for the random effects alignment function based on updated merit focus weights provided by ADEPT 2026-01-21
+
 ## 0.5.10
 
 ### Added
@@ -27,7 +37,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Added support for raw text system prompts for the pipeline baseline ADM
 * Added tagging ADM configs
 
- 
+
 ### Changed
 
 * Refactored ICL selection strategies to reduce duplication; factored out similarity strategies
@@ -36,7 +46,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Removed deprecated `tags` property from Dialog
 * Removed lots of old pre-Hydra/Outlines code
 * Made comparative regression reasoning length configurable
- 
+
 ### Fixed
 
 * Removed non-determinism from midpoint alignment functions